modified the models.py
