+++ backend/API/auth.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, request, jsonify
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_login import login_user, logout_user, login_required, current_user
-from models import Users, db
+from backend.API.models.models import Users, db
 
 
 auth_bp = Blueprint("auth_bp", __name__)
